RequirementMachine: Merge RequirementMachineImpl.h into RequirementMachine.h
