+++ lib/AST/RequirementMachine/RequirementMachine.h
@@ -14,6 +14,13 @@
 #define SWIFT_REQUIREMENTMACHINE_H
 
 #include "swift/AST/GenericSignature.h"
+#include "llvm/ADT/DenseMap.h"
+#include <vector>
+
+#include "PropertyMap.h"
+#include "ProtocolGraph.h"
+#include "RewriteContext.h"
+#include "RewriteSystem.h"
 
 namespace llvm {
 class raw_ostream;
@@ -79,6 +86,35 @@ public:
   TypeDecl *lookupNestedType(Type depType, Identifier name) const;
 
   void dump(llvm::raw_ostream &out) const;
+};
+
+/// We use the PIMPL pattern to avoid creeping header dependencies.
+struct RequirementMachine::Implementation {
+  RewriteContext &Context;
+  RewriteSystem System;
+  PropertyMap Map;
+  CanGenericSignature Sig;
+  bool Complete = false;
+
+  /// All conformance access paths computed so far.
+  llvm::DenseMap<std::pair<CanType, ProtocolDecl *>,
+                 ConformanceAccessPath> ConformanceAccessPaths;
+
+  /// Conformance access paths computed during the last round. All elements
+  /// have the same length. If a conformance access path of greater length
+  /// is requested, we refill CurrentConformanceAccessPaths with all paths of
+  /// length N+1, and add them to the ConformanceAccessPaths map.
+  std::vector<std::pair<CanType, ConformanceAccessPath>>
+      CurrentConformanceAccessPaths;
+
+  explicit Implementation(RewriteContext &ctx)
+      : Context(ctx),
+        System(Context),
+        Map(Context, System.getProtocols()) {}
+  void verify(const MutableTerm &term);
+  void dump(llvm::raw_ostream &out);
+
+  MutableTerm getLongestValidPrefix(const MutableTerm &term);
 };
 
 } // end namespace rewriting
